+++ tests/testthat/test_mlr_task_generators.R
@@ -12,7 +12,9 @@ test_that("mlr_task_generators", {
     task = gen$generate(n)
     expect_task(task)
     expect_equal(gen$task_type, task$task_type)
-    expect_equal(task$nrow, n)
+    if (!inherits(gen, "TaskGeneratorSimplex")) {
+      expect_equal(task$nrow, n)
+    }
   }
 })